daemon: move single snap querying and ops to api_snaps.go
