+++ daemon/api.go
@@ -20,28 +20,17 @@
 package daemon
 
 import (
-	"context"
-	"encoding/json"
-	"errors"
 	"fmt"
 	"net/http"
 	"strings"
 
 	"github.com/gorilla/mux"
 
-	"github.com/snapcore/snapd/client"
-	"github.com/snapcore/snapd/i18n"
-	"github.com/snapcore/snapd/logger"
 	"github.com/snapcore/snapd/overlord/assertstate"
 	"github.com/snapcore/snapd/overlord/auth"
-	"github.com/snapcore/snapd/overlord/servicestate"
 	"github.com/snapcore/snapd/overlord/snapshotstate"
 	"github.com/snapcore/snapd/overlord/snapstate"
 	"github.com/snapcore/snapd/overlord/state"
-	"github.com/snapcore/snapd/progress"
-	"github.com/snapcore/snapd/sandbox"
-	"github.com/snapcore/snapd/snap"
-	"github.com/snapcore/snapd/snap/channel"
 	"github.com/snapcore/snapd/strutil"
 )
 
@@ -89,17 +78,6 @@ var api = []*Command{
 	systemRecoveryKeysCmd,
 }
 
-var (
-	// see daemon.go:canAccess for details how the access is controlled
-	snapCmd = &Command{
-		Path:     "/v2/snaps/{name}",
-		UserOK:   true,
-		PolkitOK: "io.snapcraft.snapd.manage",
-		GET:      getSnapInfo,
-		POST:     postSnap,
-	}
-)
-
 // UserFromRequest extracts user information from request and return the respective user in state, if valid
 // It requires the state to be locked
 func UserFromRequest(st *state.State, req *http.Request) (*auth.UserState, error) {
@@ -135,166 +113,12 @@ func UserFromRequest(st *state.State, req *http.Request) (*auth.UserState, error
 
 var muxVars = mux.Vars
 
-func getSnapInfo(c *Command, r *http.Request, user *auth.UserState) Response {
-	vars := muxVars(r)
-	name := vars["name"]
-
-	about, err := localSnapInfo(c.d.overlord.State(), name)
-	if err != nil {
-		if err == errNoSnap {
-			return SnapNotFound(name, err)
-		}
-
-		return InternalError("%v", err)
-	}
-
-	route := c.d.router.Get(c.Path)
-	if route == nil {
-		return InternalError("cannot find route for %q snap", name)
-	}
-
-	url, err := route.URL("name", name)
-	if err != nil {
-		return InternalError("cannot build URL for %q snap: %v", name, err)
-	}
-
-	sd := servicestate.NewStatusDecorator(progress.Null)
-
-	result := webify(mapLocal(about, sd), url.String())
-
-	return SyncResponse(result, nil)
-}
-
-func webify(result *client.Snap, resource string) *client.Snap {
-	if result.Icon == "" || strings.HasPrefix(result.Icon, "http") {
-		return result
-	}
-	result.Icon = ""
-
-	route := appIconCmd.d.router.Get(appIconCmd.Path)
-	if route != nil {
-		url, err := route.URL("name", result.Name)
-		if err == nil {
-			result.Icon = url.String()
-		}
-	}
-
-	return result
-}
-
 func getStore(c *Command) snapstate.StoreService {
 	st := c.d.overlord.State()
 	st.Lock()
 	defer st.Unlock()
 
 	return snapstate.Store(st, nil)
-}
-
-type snapRevisionOptions struct {
-	Channel  string        `json:"channel"`
-	Revision snap.Revision `json:"revision"`
-
-	CohortKey   string `json:"cohort-key"`
-	LeaveCohort bool   `json:"leave-cohort"`
-}
-
-func (ropt *snapRevisionOptions) validate() error {
-	if ropt.CohortKey != "" {
-		if ropt.LeaveCohort {
-			return fmt.Errorf("cannot specify both cohort-key and leave-cohort")
-		}
-		if !ropt.Revision.Unset() {
-			return fmt.Errorf("cannot specify both cohort-key and revision")
-		}
-	}
-
-	if ropt.Channel != "" {
-		_, err := channel.Parse(ropt.Channel, "-")
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-type snapInstruction struct {
-	progress.NullMeter
-
-	Action string `json:"action"`
-	Amend  bool   `json:"amend"`
-	snapRevisionOptions
-	DevMode          bool     `json:"devmode"`
-	JailMode         bool     `json:"jailmode"`
-	Classic          bool     `json:"classic"`
-	IgnoreValidation bool     `json:"ignore-validation"`
-	IgnoreRunning    bool     `json:"ignore-running"`
-	Unaliased        bool     `json:"unaliased"`
-	Purge            bool     `json:"purge,omitempty"`
-	Snaps            []string `json:"snaps"`
-	Users            []string `json:"users"`
-
-	// The fields below should not be unmarshalled into. Do not export them.
-	userID int
-	ctx    context.Context
-}
-
-func (inst *snapInstruction) revnoOpts() *snapstate.RevisionOptions {
-	return &snapstate.RevisionOptions{
-		Channel:     inst.Channel,
-		Revision:    inst.Revision,
-		CohortKey:   inst.CohortKey,
-		LeaveCohort: inst.LeaveCohort,
-	}
-}
-
-func (inst *snapInstruction) modeFlags() (snapstate.Flags, error) {
-	return modeFlags(inst.DevMode, inst.JailMode, inst.Classic)
-}
-
-func (inst *snapInstruction) installFlags() (snapstate.Flags, error) {
-	flags, err := inst.modeFlags()
-	if err != nil {
-		return snapstate.Flags{}, err
-	}
-	if inst.Unaliased {
-		flags.Unaliased = true
-	}
-	if inst.IgnoreRunning {
-		flags.IgnoreRunning = true
-	}
-
-	return flags, nil
-}
-
-func (inst *snapInstruction) validate() error {
-	if inst.CohortKey != "" {
-		if inst.Action != "install" && inst.Action != "refresh" && inst.Action != "switch" {
-			return fmt.Errorf("cohort-key can only be specified for install, refresh, or switch")
-		}
-	}
-	if inst.LeaveCohort {
-		if inst.Action != "refresh" && inst.Action != "switch" {
-			return fmt.Errorf("leave-cohort can only be specified for refresh or switch")
-		}
-	}
-	if inst.Action == "install" {
-		for _, snapName := range inst.Snaps {
-			// FIXME: alternatively we could simply mutate *inst
-			//        and s/ubuntu-core/core/ ?
-			if snapName == "ubuntu-core" {
-				return fmt.Errorf(`cannot install "ubuntu-core", please use "core" instead`)
-			}
-		}
-	}
-
-	return inst.snapRevisionOptions.validate()
-}
-
-type snapInstructionResult struct {
-	Summary  string
-	Affected []string
-	Tasksets []*state.TaskSet
-	Result   map[string]interface{}
 }
 
 var (
@@ -326,250 +150,6 @@ func ensureStateSoonImpl(st *state.State) {
 }
 
 var ensureStateSoon = ensureStateSoonImpl
-
-var errDevJailModeConflict = errors.New("cannot use devmode and jailmode flags together")
-var errClassicDevmodeConflict = errors.New("cannot use classic and devmode flags together")
-var errNoJailMode = errors.New("this system cannot honour the jailmode flag")
-
-func modeFlags(devMode, jailMode, classic bool) (snapstate.Flags, error) {
-	flags := snapstate.Flags{}
-	devModeOS := sandbox.ForceDevMode()
-	switch {
-	case jailMode && devModeOS:
-		return flags, errNoJailMode
-	case jailMode && devMode:
-		return flags, errDevJailModeConflict
-	case devMode && classic:
-		return flags, errClassicDevmodeConflict
-	}
-	// NOTE: jailmode and classic are allowed together. In that setting,
-	// jailmode overrides classic and the app gets regular (non-classic)
-	// confinement.
-	flags.JailMode = jailMode
-	flags.Classic = classic
-	flags.DevMode = devMode
-	return flags, nil
-}
-
-func snapInstall(inst *snapInstruction, st *state.State) (string, []*state.TaskSet, error) {
-	if len(inst.Snaps[0]) == 0 {
-		return "", nil, fmt.Errorf(i18n.G("cannot install snap with empty name"))
-	}
-
-	flags, err := inst.installFlags()
-	if err != nil {
-		return "", nil, err
-	}
-
-	var ckey string
-	if inst.CohortKey == "" {
-		logger.Noticef("Installing snap %q revision %s", inst.Snaps[0], inst.Revision)
-	} else {
-		ckey = strutil.ElliptLeft(inst.CohortKey, 10)
-		logger.Noticef("Installing snap %q from cohort %q", inst.Snaps[0], ckey)
-	}
-	tset, err := snapstateInstall(inst.ctx, st, inst.Snaps[0], inst.revnoOpts(), inst.userID, flags)
-	if err != nil {
-		return "", nil, err
-	}
-
-	msg := fmt.Sprintf(i18n.G("Install %q snap"), inst.Snaps[0])
-	if inst.Channel != "stable" && inst.Channel != "" {
-		msg += fmt.Sprintf(" from %q channel", inst.Channel)
-	}
-	if inst.CohortKey != "" {
-		msg += fmt.Sprintf(" from %q cohort", ckey)
-	}
-	return msg, []*state.TaskSet{tset}, nil
-}
-
-func snapUpdate(inst *snapInstruction, st *state.State) (string, []*state.TaskSet, error) {
-	// TODO: bail if revision is given (and != current?), *or* behave as with install --revision?
-	flags, err := inst.modeFlags()
-	if err != nil {
-		return "", nil, err
-	}
-	if inst.IgnoreValidation {
-		flags.IgnoreValidation = true
-	}
-	if inst.IgnoreRunning {
-		flags.IgnoreRunning = true
-	}
-	if inst.Amend {
-		flags.Amend = true
-	}
-
-	// we need refreshed snap-declarations to enforce refresh-control as best as we can
-	if err = assertstateRefreshSnapDeclarations(st, inst.userID); err != nil {
-		return "", nil, err
-	}
-
-	ts, err := snapstateUpdate(st, inst.Snaps[0], inst.revnoOpts(), inst.userID, flags)
-	if err != nil {
-		return "", nil, err
-	}
-
-	msg := fmt.Sprintf(i18n.G("Refresh %q snap"), inst.Snaps[0])
-	if inst.Channel != "stable" && inst.Channel != "" {
-		msg = fmt.Sprintf(i18n.G("Refresh %q snap from %q channel"), inst.Snaps[0], inst.Channel)
-	}
-
-	return msg, []*state.TaskSet{ts}, nil
-}
-
-func snapRemove(inst *snapInstruction, st *state.State) (string, []*state.TaskSet, error) {
-	ts, err := snapstate.Remove(st, inst.Snaps[0], inst.Revision, &snapstate.RemoveFlags{Purge: inst.Purge})
-	if err != nil {
-		return "", nil, err
-	}
-
-	msg := fmt.Sprintf(i18n.G("Remove %q snap"), inst.Snaps[0])
-	return msg, []*state.TaskSet{ts}, nil
-}
-
-func snapRevert(inst *snapInstruction, st *state.State) (string, []*state.TaskSet, error) {
-	var ts *state.TaskSet
-
-	flags, err := inst.modeFlags()
-	if err != nil {
-		return "", nil, err
-	}
-
-	if inst.Revision.Unset() {
-		ts, err = snapstateRevert(st, inst.Snaps[0], flags)
-	} else {
-		ts, err = snapstateRevertToRevision(st, inst.Snaps[0], inst.Revision, flags)
-	}
-	if err != nil {
-		return "", nil, err
-	}
-
-	msg := fmt.Sprintf(i18n.G("Revert %q snap"), inst.Snaps[0])
-	return msg, []*state.TaskSet{ts}, nil
-}
-
-func snapEnable(inst *snapInstruction, st *state.State) (string, []*state.TaskSet, error) {
-	if !inst.Revision.Unset() {
-		return "", nil, errors.New("enable takes no revision")
-	}
-	ts, err := snapstate.Enable(st, inst.Snaps[0])
-	if err != nil {
-		return "", nil, err
-	}
-
-	msg := fmt.Sprintf(i18n.G("Enable %q snap"), inst.Snaps[0])
-	return msg, []*state.TaskSet{ts}, nil
-}
-
-func snapDisable(inst *snapInstruction, st *state.State) (string, []*state.TaskSet, error) {
-	if !inst.Revision.Unset() {
-		return "", nil, errors.New("disable takes no revision")
-	}
-	ts, err := snapstate.Disable(st, inst.Snaps[0])
-	if err != nil {
-		return "", nil, err
-	}
-
-	msg := fmt.Sprintf(i18n.G("Disable %q snap"), inst.Snaps[0])
-	return msg, []*state.TaskSet{ts}, nil
-}
-
-func snapSwitch(inst *snapInstruction, st *state.State) (string, []*state.TaskSet, error) {
-	if !inst.Revision.Unset() {
-		return "", nil, errors.New("switch takes no revision")
-	}
-	ts, err := snapstateSwitch(st, inst.Snaps[0], inst.revnoOpts())
-	if err != nil {
-		return "", nil, err
-	}
-
-	var msg string
-	switch {
-	case inst.LeaveCohort && inst.Channel != "":
-		msg = fmt.Sprintf(i18n.G("Switch %q snap to channel %q and away from cohort"), inst.Snaps[0], inst.Channel)
-	case inst.LeaveCohort:
-		msg = fmt.Sprintf(i18n.G("Switch %q snap away from cohort"), inst.Snaps[0])
-	case inst.CohortKey == "" && inst.Channel != "":
-		msg = fmt.Sprintf(i18n.G("Switch %q snap to channel %q"), inst.Snaps[0], inst.Channel)
-	case inst.CohortKey != "" && inst.Channel == "":
-		msg = fmt.Sprintf(i18n.G("Switch %q snap to cohort %q"), inst.Snaps[0], strutil.ElliptLeft(inst.CohortKey, 10))
-	default:
-		msg = fmt.Sprintf(i18n.G("Switch %q snap to channel %q and cohort %q"), inst.Snaps[0], inst.Channel, strutil.ElliptLeft(inst.CohortKey, 10))
-	}
-	return msg, []*state.TaskSet{ts}, nil
-}
-
-type snapActionFunc func(*snapInstruction, *state.State) (string, []*state.TaskSet, error)
-
-var snapInstructionDispTable = map[string]snapActionFunc{
-	"install": snapInstall,
-	"refresh": snapUpdate,
-	"remove":  snapRemove,
-	"revert":  snapRevert,
-	"enable":  snapEnable,
-	"disable": snapDisable,
-	"switch":  snapSwitch,
-}
-
-func (inst *snapInstruction) dispatch() snapActionFunc {
-	if len(inst.Snaps) != 1 {
-		logger.Panicf("dispatch only handles single-snap ops; got %d", len(inst.Snaps))
-	}
-	return snapInstructionDispTable[inst.Action]
-}
-
-func (inst *snapInstruction) errToResponse(err error) Response {
-	if len(inst.Snaps) == 0 {
-		return errToResponse(err, nil, BadRequest, "cannot %s: %v", inst.Action)
-	}
-
-	return errToResponse(err, inst.Snaps, BadRequest, "cannot %s %s: %v", inst.Action, strutil.Quoted(inst.Snaps))
-}
-
-func postSnap(c *Command, r *http.Request, user *auth.UserState) Response {
-	route := c.d.router.Get(stateChangeCmd.Path)
-	if route == nil {
-		return InternalError("cannot find route for change")
-	}
-
-	decoder := json.NewDecoder(r.Body)
-	var inst snapInstruction
-	if err := decoder.Decode(&inst); err != nil {
-		return BadRequest("cannot decode request body into snap instruction: %v", err)
-	}
-	inst.ctx = r.Context()
-
-	state := c.d.overlord.State()
-	state.Lock()
-	defer state.Unlock()
-
-	if user != nil {
-		inst.userID = user.ID
-	}
-
-	vars := muxVars(r)
-	inst.Snaps = []string{vars["name"]}
-
-	if err := inst.validate(); err != nil {
-		return BadRequest("%s", err)
-	}
-
-	impl := inst.dispatch()
-	if impl == nil {
-		return BadRequest("unknown action %s", inst.Action)
-	}
-
-	msg, tsets, err := impl(&inst, state)
-	if err != nil {
-		return inst.errToResponse(err)
-	}
-
-	chg := newChange(state, inst.Action+"-snap", msg, tsets, inst.Snaps)
-
-	ensureStateSoon(state)
-
-	return AsyncResponse(nil, &Meta{Change: chg.ID()})
-}
 
 func newChange(st *state.State, kind, summary string, tsets []*state.TaskSet, snapNames []string) *state.Change {
 	chg := st.NewChange(kind, summary)
